test: refactoring to use environment
